refactor: separated product dependencies into repos and usecases
