+++ products/infrastructure/fastapi/dependencies/usecases.py
@@ -1,7 +1,7 @@
-from functools import lru_cache
-
 from fastapi import Depends
 
+from activity.domain.record_repository import RecordRepository
+from activity.infrastructure.fastapi.dependencies import get_record_repository
 from core.infrastructure.fastapi.dependencies import get_mailer
 from emails.domain.emailer import Emailer
 from products.application.create_product import CreateProductUseCase
@@ -12,19 +12,8 @@ from products.application.register_arrival import RegisterArrivalUsecase
 from products.application.register_sell import RegisterSaleUsecase
 from products.application.search_by_name import SearchProductsByNameUsecase
 from products.domain.product_repository import ProductRepository
-from products.infrastructure.sqlite_product_repository import SQLiteProductRepository
-from products.infrastructure.supabase_product_repository import (
-    SupabaseProductRepository,
-)
-from shared.infrastructure.env import ENV
+from products.infrastructure.fastapi.dependencies.repos import get_product_repository
 from shared.infrastructure.basic_logger import basic_logger
-
-
-@lru_cache
-def get_product_repository() -> ProductRepository:
-    if ENV.SQLITE_PATH:
-        return SQLiteProductRepository(ENV.SQLITE_PATH)
-    return SupabaseProductRepository()
 
 
 def get_create_product_usecase(
@@ -36,15 +25,19 @@ def get_create_product_usecase(
 
 def get_register_sale_usecase(
     product_repository: ProductRepository = Depends(get_product_repository),
+    record_repository: RecordRepository = Depends(get_record_repository),
     mailer: Emailer = Depends(get_mailer),
 ) -> RegisterSaleUsecase:
-    return RegisterSaleUsecase(product_repository, mailer, basic_logger)
+    return RegisterSaleUsecase(
+        product_repository, record_repository, mailer, basic_logger
+    )
 
 
 def get_register_arrival_usecase(
     product_repository: ProductRepository = Depends(get_product_repository),
+    record_repository: RecordRepository = Depends(get_record_repository),
 ) -> RegisterArrivalUsecase:
-    return RegisterArrivalUsecase(basic_logger, product_repository)
+    return RegisterArrivalUsecase(basic_logger, product_repository, record_repository)
 
 
 def get_all_products_usecase(
@@ -60,9 +53,10 @@ def get_product_by_id_usecase(
 
 
 def get_delete_product_usecase(
-    repo: ProductRepository = Depends(get_product_repository),
+    product_repository: ProductRepository = Depends(get_product_repository),
+    record_repository: RecordRepository = Depends(get_record_repository),
 ) -> DeleteProductByIdUsecase:
-    return DeleteProductByIdUsecase(basic_logger, repo)
+    return DeleteProductByIdUsecase(basic_logger, product_repository, record_repository)
 
 
 def get_search_products_by_name_usecase(
